requirements.txt
catkin_pkg
numpy
pytorch-ignite
matplotlib
tensorboard
tqdm
torch
imgaug